move static accounts compile into transaction-messages
