+++ packages/transaction-messages/src/compile/index.ts
@@ -5,3 +5,4 @@ export * from './address-table-lookups';
 export * from './header';
 export * from './instructions';
 export * from './lifetime-token';
+export * from './static-accounts';
